Implement example models for mujoco
Hello.xml implements a simple floating box with 6 DOF. Defines a plane fixed to the world,
a light to better illuminate objects and cast shadows. The built-in opengl visualizer runs
the simulation and makes the box observe gravitational forces.

Example.xml implements a 7 DOF arm that holds a string with a cylinder attached at the other end.
The string is implemented as a tendon with length limits. There is ball joint at the shoulder
and pairs of hinge joints at the elbow and wrist. The box inside the cylinder indicates a free “joint”.
The outer body element in the XML is the required worldbody.
Note that using multiple joints between two bodies does not require creating dummy bodies.