walk.py:
upload scans all the github links of all folders in the
path provided (or ./ by default). Collects and
uploads all of these to a file. (useful to replicate
my other_git codebases)

download does the opposite and collects all of the structure
locally, including folder names

method : this should be a tree search (either breadth first or
depth first) where a leaf node is one that yields a valid
string for git remote -v

to make this simpler:

1. We let os.walk do the traversal for us in depth first search

2. Only when we see a .git in the current folder, we do a git remote -v and capture it
(root_folder (relative) + dirpath and cmd_output as JSON fields so they can be
re-used for reconstructing)

3. Technically (2) is a leaf. If we find a way to prune the processing in os.walk
we do it, else we leave it and let it continue (there could be corner cases of
3rd party git installs under a .git (with a different .git) that a prune would
miss capturing).